+++ modules/core/src/main/java/net/ashald/envfile/providers/runconfig/RunConfigEnvVarsProvider.java
@@ -4,10 +4,12 @@ import net.ashald.envfile.AbstractEnvVarsProvider;
 import org.jetbrains.annotations.NotNull;
 
 import java.io.File;
+import java.io.InputStream;
 import java.util.HashMap;
 import java.util.Map;
 
 public class RunConfigEnvVarsProvider extends AbstractEnvVarsProvider {
+    public static final String PARSER_ID = "runconfig";
 
     public RunConfigEnvVarsProvider(boolean shouldSubstituteEnvVar) {
         super(shouldSubstituteEnvVar);
@@ -15,7 +17,7 @@ public class RunConfigEnvVarsProvider extends AbstractEnvVarsProvider {
 
     @NotNull
     @Override
-    protected Map<String, String> getEnvVars(@NotNull Map<String, String> runConfigEnv, String path) {
+    protected Map<String, String> getEnvVars(@NotNull Map<String, String> runConfigEnv, InputStream content) {
         return new HashMap<>(runConfigEnv);
     }