+++ modules/platform/src/main/java/net/ashald/envfile/platform/ui/EnvFileConfigurationPanel.java
@@ -15,7 +15,6 @@ import com.intellij.openapi.ui.popup.ListPopup;
 import com.intellij.openapi.util.Conditions;
 import com.intellij.openapi.vfs.VirtualFile;
 import com.intellij.ui.AnActionButton;
-import com.intellij.ui.AnActionButtonRunnable;
 import com.intellij.ui.AnActionButtonUpdater;
 import com.intellij.ui.BooleanTableCellRenderer;
 import com.intellij.ui.ToolbarDecorator;
@@ -28,8 +27,10 @@ import net.ashald.envfile.platform.EnvFileEntry;
 import net.ashald.envfile.platform.EnvVarsProviderExtension;
 import net.ashald.envfile.platform.EnvFileSettings;
 import net.ashald.envfile.platform.ui.table.EnvFileIsActiveColumnInfo;
+import net.ashald.envfile.platform.ui.table.EnvFileIsExecutableColumnInfo;
 import net.ashald.envfile.platform.ui.table.EnvFilePathColumnInfo;
 import net.ashald.envfile.platform.ui.table.EnvFileTypeColumnInfo;
+import net.ashald.envfile.providers.runconfig.RunConfigEnvVarsProvider;
 
 import javax.swing.BoxLayout;
 import javax.swing.JPanel;
@@ -47,7 +48,7 @@ import java.util.LinkedList;
 
 class EnvFileConfigurationPanel<T extends RunConfigurationBase<?>> extends JPanel {
     private static final int MAX_RECENT = 5;
-    private static final LinkedList<EnvFileEntry> recent = new LinkedList<EnvFileEntry>();
+    private static final LinkedList<EnvFileEntry> RECENT = new LinkedList<>();
     private final RunConfigurationBase<?> runConfig;
 
     private final JCheckBox useEnvFileCheckBox;
@@ -62,18 +63,20 @@ class EnvFileConfigurationPanel<T extends RunConfigurationBase<?>> extends JPane
         runConfig = configuration;
 
         // Define Model
-        ColumnInfo<EnvFileEntry, Boolean> IS_ACTIVE = new EnvFileIsActiveColumnInfo();
-        ColumnInfo<EnvFileEntry, String> FILE = new EnvFilePathColumnInfo();
-        ColumnInfo<EnvFileEntry, EnvFileEntry> TYPE = new EnvFileTypeColumnInfo();
+        ColumnInfo<EnvFileEntry, Boolean> columnIsActive = new EnvFileIsActiveColumnInfo();
+        ColumnInfo<EnvFileEntry, Boolean> columnIsExecutable = new EnvFileIsExecutableColumnInfo();
+        ColumnInfo<EnvFileEntry, String> columnFile = new EnvFilePathColumnInfo();
+        ColumnInfo<EnvFileEntry, EnvFileEntry> columnType = new EnvFileTypeColumnInfo();
 
-        envFilesModel = new ListTableModel<>(IS_ACTIVE, FILE, TYPE);
+        envFilesModel = new ListTableModel<>(columnIsActive, columnIsExecutable, columnFile, columnType);
 
         // Create Table
         envFilesTable = new TableView<>(envFilesModel);
         envFilesTable.getEmptyText().setText("No environment variables files selected");
 
-        setUpColumnWidth(envFilesTable, 0, IS_ACTIVE, 20);
-        setUpColumnWidth(envFilesTable, 2, TYPE, 50);
+        setUpColumnWidth(envFilesTable, 0, columnIsActive, 20);
+        setUpColumnWidth(envFilesTable, 1, columnIsExecutable, 20);
+        setUpColumnWidth(envFilesTable, 3, columnType, 50);
 
         envFilesTable.setColumnSelectionAllowed(false);
         envFilesTable.setShowGrid(false);
@@ -101,40 +104,35 @@ class EnvFileConfigurationPanel<T extends RunConfigurationBase<?>> extends JPane
         experimentalIntegrationsCheckBox = new JCheckBox("Enable experimental integrations (e.g. Gradle) - may break any time!");
 
         // TODO: come up with a generic approach for this
-        envFilesModel.addRow(new EnvFileEntry(runConfig, "runconfig", null, true, substituteEnvVarsCheckBox.isSelected()));
+        envFilesModel.addRow(
+                EnvFileEntry.builder()
+                        .runConfig(runConfig)
+                        .parserId(RunConfigEnvVarsProvider.PARSER_ID)
+                        .enabled(true)
+                        .substitutionEnabled(true)
+                        .executable(false)
+                        .build()
+        );
 
         // Create Toolbar - Add/Remove/Move actions
         final ToolbarDecorator envFilesTableDecorator = ToolbarDecorator.createDecorator(envFilesTable);
 
-        final AnActionButtonUpdater updater = new AnActionButtonUpdater() {
-            @Override
-            public boolean isEnabled(AnActionEvent e) {
-                return useEnvFileCheckBox.isSelected();
-            }
-        };
+        final AnActionButtonUpdater updater = e -> useEnvFileCheckBox.isSelected();
 
         envFilesTableDecorator
-                .setAddAction(new AnActionButtonRunnable() {
-            @Override
-            public void run(AnActionButton button) {
-                doAddAction(button, envFilesTable, envFilesModel);
-            }
-        })
+                .setAddAction(button -> doAddAction(button, envFilesTable, envFilesModel))
                 .setAddActionUpdater(updater)
-                .setRemoveActionUpdater(new AnActionButtonUpdater() {
-                    @Override
-                    public boolean isEnabled(AnActionEvent e) {
-                        boolean allEditable = true;
+                .setRemoveActionUpdater(e -> {
+                    boolean allEditable = true;
 
-                        for (EnvFileEntry entry : envFilesTable.getSelectedObjects()) {
-                            if (!entry.isEditable()) {
-                                allEditable = false;
-                                break;
-                            }
+                    for (EnvFileEntry entry : envFilesTable.getSelectedObjects()) {
+                        if (!entry.isEditable()) {
+                            allEditable = false;
+                            break;
                         }
-
-                        return updater.isEnabled(e) && envFilesTable.getSelectedRowCount() >= 1 && allEditable;
                     }
+
+                    return updater.isEnabled(e) && envFilesTable.getSelectedRowCount() >= 1 && allEditable;
                 });
 
         JPanel optionsPanel = new JPanel();
@@ -207,28 +205,36 @@ class EnvFileConfigurationPanel<T extends RunConfigurationBase<?>> extends JPane
                         }
 
                         ArrayList<EnvFileEntry> newList = new ArrayList<EnvFileEntry>(model.getItems());
-                        final EnvFileEntry newOptions = new EnvFileEntry(runConfig, extension.getId(), selectedPath, true, substituteEnvVarsCheckBox.isSelected());
+                        final EnvFileEntry newOptions = EnvFileEntry.builder()
+                                .runConfig(runConfig)
+                                .parserId(extension.getId())
+                                .path(selectedPath)
+                                .enabled(true)
+                                .substitutionEnabled(substituteEnvVarsCheckBox.isSelected())
+                                .executable(false)
+                                .build();
+
                         newList.add(newOptions);
                         model.setItems(newList);
                         int index = model.getRowCount() - 1;
                         model.fireTableRowsInserted(index, index);
                         table.setRowSelectionInterval(index, index);
 
-                        synchronized (recent) {
-                            recent.remove(newOptions);
-                            recent.addFirst(newOptions);
-                            if (recent.size() > MAX_RECENT) recent.removeLast();
+                        synchronized (RECENT) {
+                            RECENT.remove(newOptions);
+                            RECENT.addFirst(newOptions);
+                            if (RECENT.size() > MAX_RECENT) RECENT.removeLast();
                         }
                     }
                 }
             };
             actionGroup.add(anAction);
         }
-        synchronized (recent) {
-            if (!recent.isEmpty()) {
+        synchronized (RECENT) {
+            if (!RECENT.isEmpty()) {
                 actionGroup.addSeparator("Recent");
 
-                for (final EnvFileEntry entry : recent) {
+                for (final EnvFileEntry entry : RECENT) {
                     String title = String.format("%s -> %s", entry.getTypeTitle(), entry.getPath());
                     String shortTitle = title.length() < 81 ? title : title.replaceFirst("(.{39}).+(.{39})", "$1...$2");
                     AnAction anAction = new AnAction(shortTitle, title, null) {
@@ -240,9 +246,9 @@ class EnvFileConfigurationPanel<T extends RunConfigurationBase<?>> extends JPane
                             int index = model.getRowCount() - 1;
                             model.fireTableRowsInserted(index, index);
                             table.setRowSelectionInterval(index, index);
-                            synchronized (recent) {
-                                recent.remove(entry);
-                                recent.addFirst(entry);
+                            synchronized (RECENT) {
+                                RECENT.remove(entry);
+                                RECENT.addFirst(entry);
                             }
                         }
                     };
@@ -264,28 +270,28 @@ class EnvFileConfigurationPanel<T extends RunConfigurationBase<?>> extends JPane
     }
 
     EnvFileSettings getState() {
-        return new EnvFileSettings(
-                useEnvFileCheckBox.isSelected(),
-                substituteEnvVarsCheckBox.isSelected(),
-                supportPathMacroCheckBox.isSelected(),
-                envFilesModel.getItems(),
-                ignoreMissingCheckBox.isSelected(),
-                experimentalIntegrationsCheckBox.isSelected()
-        );
+        return EnvFileSettings.builder()
+                .pluginEnabled(useEnvFileCheckBox.isSelected())
+                .envVarsSubstitutionEnabled(substituteEnvVarsCheckBox.isSelected())
+                .pathMacroSupported(supportPathMacroCheckBox.isSelected())
+                .ignoreMissing(ignoreMissingCheckBox.isSelected())
+                .enableExperimentalIntegrations(experimentalIntegrationsCheckBox.isSelected())
+                .entries(envFilesModel.getItems())
+                .build();
     }
 
     void setState(EnvFileSettings state) {
-        useEnvFileCheckBox.setSelected(state.isEnabled());
+        useEnvFileCheckBox.setSelected(state.isPluginEnabledEnabled());
         substituteEnvVarsCheckBox.setSelected(state.isSubstituteEnvVarsEnabled());
         supportPathMacroCheckBox.setSelected(state.isPathMacroSupported());
         ignoreMissingCheckBox.setSelected(state.isIgnoreMissing());
         experimentalIntegrationsCheckBox.setSelected(state.isEnableExperimentalIntegrations());
 
-        envFilesTable.setEnabled(state.isEnabled());
-        substituteEnvVarsCheckBox.setEnabled(state.isEnabled());
-        supportPathMacroCheckBox.setEnabled(state.isEnabled());
-        ignoreMissingCheckBox.setEnabled(state.isEnabled());
-        experimentalIntegrationsCheckBox.setEnabled(state.isEnabled());
+        envFilesTable.setEnabled(state.isPluginEnabledEnabled());
+        substituteEnvVarsCheckBox.setEnabled(state.isPluginEnabledEnabled());
+        supportPathMacroCheckBox.setEnabled(state.isPluginEnabledEnabled());
+        ignoreMissingCheckBox.setEnabled(state.isPluginEnabledEnabled());
+        experimentalIntegrationsCheckBox.setEnabled(state.isPluginEnabledEnabled());
 
         envFilesModel.setItems(new ArrayList<>(state.getEntries()));
     }