+++ modules/platform/src/main/java/net/ashald/envfile/platform/ui/table/EnvFileIsExecutableColumnInfo.java
@@ -1,5 +1,6 @@
 package net.ashald.envfile.platform.ui.table;
 
+import com.intellij.openapi.util.NlsContexts;
 import com.intellij.ui.BooleanTableCellRenderer;
 import com.intellij.util.ui.ColumnInfo;
 import net.ashald.envfile.platform.EnvFileEntry;
@@ -10,15 +11,15 @@ import javax.swing.JTable;
 import javax.swing.table.TableCellRenderer;
 import java.awt.Component;
 
-public class EnvFileIsActiveColumnInfo extends ColumnInfo<EnvFileEntry, Boolean> {
-    public EnvFileIsActiveColumnInfo() {
-        super("Enabled");
+public class EnvFileIsExecutableColumnInfo extends ColumnInfo<EnvFileEntry, Boolean> {
+    public EnvFileIsExecutableColumnInfo() {
+        super("Executable");
     }
 
     @Nullable
     @Override
     public Boolean valueOf(EnvFileEntry envFileEntry) {
-        return envFileEntry.isEnabled();
+        return envFileEntry.getExecutable();
     }
 
     @Override
@@ -28,12 +29,12 @@ public class EnvFileIsActiveColumnInfo extends ColumnInfo<EnvFileEntry, Boolean>
 
     @Override
     public void setValue(EnvFileEntry element, Boolean checked) {
-        element.setEnabled(checked);
+        element.setExecutable(checked);
     }
 
     @Override
     public boolean isCellEditable(EnvFileEntry envFileEntry) {
-        return envFileEntry.isEditable();
+        return envFileEntry.getPath() != null;
     }
 
     @Nullable
@@ -42,16 +43,19 @@ public class EnvFileIsActiveColumnInfo extends ColumnInfo<EnvFileEntry, Boolean>
         return new BooleanTableCellRenderer() {
             @NotNull
             @Override
-            public Component getTableCellRendererComponent(
-                    @NotNull JTable table,
-                    Object value,
-                    boolean isSelected,
-                    boolean hasFocus,
-                    int row,
-                    int column
-            ) {
+            public Component getTableCellRendererComponent(@NotNull JTable table,
+                                                           Object value,
+                                                           boolean isSelected,
+                                                           boolean hasFocus,
+                                                           int row,
+                                                           int column) {
                 return super.getTableCellRendererComponent(table, value, isSelected, hasFocus, row, column);
             }
         };
+    }
+
+    @Override
+    public @NlsContexts.Tooltip @Nullable String getTooltipText() {
+        return "Execute given file and parse content from standard output";
     }
 }